+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Latest
 * Create separate Country model from Location.
 * Create separate Region (state) model from Location.
 * Create separate SubRegion (county) model from Location.
+* Add Area model for defining groups of Locations.
 
 1.0.3 (2025-03-20)
 ------------------